refactor(aigateway): extract usage taxonomy plugin
Refs: OME-303
