+++ apps/aigateway/src/aigateway/plugins/taxonomy/money.py
