+++ apps/aigateway/src/aigateway/plugins/taxonomy/render.py
@@ -6,9 +6,9 @@ import json
 from collections.abc import Sequence
 from typing import Any, Literal
 
-from ._collector import RequestAccountingCollector
-from ._money import sum_amounts
-from ._types import (
+from .collector import RequestAccountingCollector
+from .money import sum_amounts
+from .types import (
     SCHEMA_REQUEST_ECONOMICS,
     SCHEMA_USAGE_ACCOUNTING,
     CacheReference,
