+++ apps/aigateway/tests/unit/test_plugin_loader.py
@@ -2,6 +2,9 @@
 
 from __future__ import annotations
 
+import logging
+import sys
+
 from aigateway.core.loader import load_plugins
 from aigateway.core.registry import ProviderRegistry
 
@@ -58,3 +61,58 @@ def test_loader_discovers_huggingface_provider() -> None:
     for m in models:
         assert m.litellm_params["model"].startswith("huggingface/")
         assert m.litellm_params["api_base"] == "https://router.huggingface.co/v1"
+
+
+def test_loader_silently_ignores_non_provider_plugin_packages(caplog) -> None:
+    reg = ProviderRegistry()
+
+    with caplog.at_level(logging.WARNING, logger="aigateway.core.loader"):
+        load_plugins(reg)
+
+    assert not [record for record in caplog.records if "plugins.taxonomy" in record.getMessage()]
+
+
+def test_loader_warns_when_a_provider_dependency_is_unavailable(
+    tmp_path, monkeypatch, caplog
+) -> None:
+    package = tmp_path / "test_plugins"
+    provider = package / "broken_provider"
+    package.mkdir()
+    provider.mkdir()
+    (package / "__init__.py").write_text("")
+    (provider / "__init__.py").write_text("")
+    (provider / "plugin.py").write_text("import dependency_that_does_not_exist\n")
+    monkeypatch.syspath_prepend(str(tmp_path))
+
+    with caplog.at_level(logging.WARNING, logger="aigateway.core.loader"):
+        load_plugins(ProviderRegistry(), package="test_plugins")
+
+    assert (
+        "provider plugin test_plugins.broken_provider has an unavailable dependency" in caplog.text
+    )
+    for name in tuple(sys.modules):
+        if name == "test_plugins" or name.startswith("test_plugins."):
+            sys.modules.pop(name)
+
+
+def test_loader_warns_when_a_provider_package_dependency_is_unavailable(
+    tmp_path, monkeypatch, caplog
+) -> None:
+    package = tmp_path / "test_plugins"
+    provider = package / "broken_provider"
+    package.mkdir()
+    provider.mkdir()
+    (package / "__init__.py").write_text("")
+    (provider / "__init__.py").write_text("import dependency_that_does_not_exist\n")
+    (provider / "plugin.py").write_text("")
+    monkeypatch.syspath_prepend(str(tmp_path))
+
+    with caplog.at_level(logging.WARNING, logger="aigateway.core.loader"):
+        load_plugins(ProviderRegistry(), package="test_plugins")
+
+    assert (
+        "provider plugin test_plugins.broken_provider has an unavailable dependency" in caplog.text
+    )
+    for name in tuple(sys.modules):
+        if name == "test_plugins" or name.startswith("test_plugins."):
+            sys.modules.pop(name)
